[JAVA-46836] Moved code from hibernate-exceptions to hibernate-exceptions-2
