+++ persistence-modules/hibernate-exceptions-2/src/main/java/com/baeldung/hibernate/entitynotfoundexception/User.java
