+++ persistence-modules/hibernate-exceptions-2/src/main/java/com/baeldung/hibernate/exception/persistentobject/entity/Author.java
@@ -1,15 +1,10 @@
 package com.baeldung.hibernate.exception.persistentobject.entity;
 
-import java.util.List;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-
+import jakarta.persistence.*;
 import org.hibernate.annotations.Cascade;
 import org.hibernate.annotations.CascadeType;
+
+import java.util.List;
 
 @Entity
 public class Author {
