+++ persistence-modules/hibernate-exceptions-2/src/main/java/com/baeldung/hibernate/exception/persistentobject/entity/Book.java
@@ -1,10 +1,6 @@
 package com.baeldung.hibernate.exception.persistentobject.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 
 @Entity
 public class Book {
