+++ persistence-modules/hibernate-exceptions-2/src/main/resources/META-INF/persistence.xml
